Make each parsing test pass/fail separately.
