+++ Cargo.toml
@@ -11,6 +11,18 @@ readme = "README.md"
 keywords = ["url", "parser"]
 license = "MIT/Apache-2.0"
 
+[[test]] name = "format"
+[[test]] name = "form_urlencoded"
+[[test]] name = "idna"
+[[test]] name = "punycode"
+[[test]] name = "tests"
+[[test]]
+name = "wpt"
+harness = false
+
+[dev-dependencies]
+rustc-test = "0.1"
+
 [features]
 query_encoding = ["encoding"]
 serde_serialization = ["serde"]
